Move portlets stuff to subpackage
