+++ src/plone/restapi/serializer/configure.zcml
@@ -3,6 +3,8 @@
     xmlns:zcml="http://namespaces.zope.org/zcml"
     i18n_domain="plone.restapi">
 
+    <include package=".portlets" />
+
     <adapter factory=".site.SerializeSiteRootToJson" />
     <adapter factory=".dxcontent.SerializeToJson" />
     <adapter factory=".dxcontent.SerializeFolderToJson" />
@@ -82,13 +84,6 @@
 
     <adapter factory=".discussion.ConversationSerializer" />
     <adapter factory=".discussion.CommentSerializer" />
-
-    <adapter factory=".portlet.PortletManagerSummarySerializer" />
-    <adapter factory=".portlet.PortletManagerSerializer" />
-    <adapter factory=".portlet.PortletSerializer" />
-    <adapter factory=".portlet.NavigationPortletSerializer" name="portlets.Navigation"/>
-    <adapter factory=".portlet.DefaultPortletFieldSerializer" />
-    <adapter factory=".portlet.RichttextPortletFieldSerializer" />
 
     <configure zcml:condition="installed plone.tiles">
         <adapter factory=".tile.TileSummarySerializeToJson" />
